+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -104,6 +104,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="BlastGeneratorProto.cs" />
     <Compile Include="BlastLayer.cs" />
+    <Compile Include="BlastTarget.cs" />
     <Compile Include="BlastTools.cs" />
     <Compile Include="BlastUnit.cs" />
     <Compile Include="ComboBoxItem.cs" />
